fix: reduce absurdly large password hash size
Previously, the password hashes generated would be 584 bytes (512 + 64 +
8) which is far larger than necessary for security. As a result they
would unnecessarily bloat ledger databases.

Cause of the bug was a misinterpretation on my part of the crypto.pbkdf2
function. I thought it expected the key length in bits, but it expects
it in bytes. (In the example a keylen of 512 is used with SHA-512, so I
think the documentation writer made the same error.)

Since a smaller output size also causes a speed-up, I'm increasing the
default iteration count to compensate.

In this commit I'm choosing sane values for salt and hash size that lead
to an overall password hash size of 40 bytes (instead of 548).

Also note that since the salt size, hash size and iteration count are
parsed from existing password hash database entries, this change is
fully backwards compatible.